punctuation: Support artist credits
- Insert a "Guess punctuation" button into the artist credit bubble
- Trigger a different event for artist credit edits ('blur' instead of
  'change'), otherwise MBS does not know about them

The bubble does not exist in the DOM until it has been opened for the first
time, so we have to listen for clicks on all buttons which could open it.
After that, the same instance is used for different ACs and we also have
to take care of highlights that are left over from the previous AC edit.