api: Stop wrapping JSON responses with MB.entity()
This was necessary to share the entity cache with MBS for the old
relationship editor, but breaks the new one during submission.
While adding a relationship to the fetched target entity looks like it
has worked in the UI, entering the edits fails silently with
"Uncaught Error: Only plain objects and arrays can be converted into JSON"
because MB.entity adds function properties to the API data.