Create import script for ARD Audiothek broadcast releases
First version which only runs on initial page load and can't detect
navigation inside the single page application (SPA).
The layout is still a bit rough but the features are almost on par with
the other ARD radio drama import script. Only the artists are missing
from the API data, so we prefill it with [unknown] to allow the user to
map it to an MBID (temporarily, for a few consecutive imports).